sandbox: Move command execution into its own package.
