+++ ledger/sandbox/src/main/scala/com/digitalasset/platform/apiserver/execution/CommandExecutorImpl.scala
@@ -1,10 +1,11 @@
 // Copyright (c) 2020 Digital Asset (Switzerland) GmbH and/or its affiliates. All rights reserved.
 // SPDX-License-Identifier: Apache-2.0
 
-package com.daml.platform.apiserver
+package com.daml.platform.apiserver.execution
 
 import java.util.concurrent.ConcurrentHashMap
 
+import com.daml.ledger.api.domain.{Commands => ApiCommands}
 import com.daml.ledger.participant.state.v1.{SubmitterInfo, TransactionMeta}
 import com.daml.lf.command._
 import com.daml.lf.crypto
@@ -26,7 +27,6 @@ import com.daml.lf.transaction.Node.GlobalKey
 import com.daml.lf.transaction.Transaction.{Value => TxValue}
 import com.daml.lf.value.Value
 import com.daml.lf.value.Value.AbsoluteContractId
-import com.daml.ledger.api.domain.{Commands => ApiCommands}
 import com.daml.platform.store.ErrorCause
 import scalaz.syntax.tag._
