+++ ledger/sandbox/src/main/scala/com/digitalasset/platform/apiserver/execution/LedgerTimeHelper.scala
@@ -1,16 +1,16 @@
 // Copyright (c) 2020 Digital Asset (Switzerland) GmbH and/or its affiliates. All rights reserved.
 // SPDX-License-Identifier: Apache-2.0
 
-package com.daml.platform.apiserver
+package com.daml.platform.apiserver.execution
 
 import java.time.Instant
 
+import com.daml.ledger.api.domain.{Commands => ApiCommands}
 import com.daml.ledger.participant.state.index.v2.ContractStore
 import com.daml.lf.crypto
 import com.daml.lf.data.Time
 import com.daml.lf.transaction.Transaction
 import com.daml.lf.value.Value.AbsoluteContractId
-import com.daml.ledger.api.domain.{Commands => ApiCommands}
 import com.daml.logging.{ContextualizedLogger, LoggingContext}
 import com.daml.platform.store.ErrorCause
